+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/RelationshipOp.java
@@ -14,5 +14,6 @@ public enum RelationshipOp {
     DELETE,
     UPSERT,
     REPLACE,
-    REMOVE
+    REMOVE,
+    UPDATE
 }