+++ changelog.md
@@ -1,10 +1,12 @@
 # Change Log
-
 ## 4.0-beta-2
 **Fixes**
  * Root collection loads now push down security filter predicates.
  * Avoid throwing exceptions that must be handled by the containing application, instead throw exceptions that will be handled directly within Elide.
  * Restore OnCreatePreSecurity lifecycle hook to occur after fields are populated.
+
+**Features**
+ * Added UPDATE operation for GraphQL.
 
 ## 4.0-beta-1
 **Features**